Fix incorrect dates at range ends
Because of a mistake, the end date of time ranges as shown in the
dashboard was not copied by value but by reference. As an effect,
displayed dates would be repeatedly decremented by one second every time
users hovered data points. For this reason, displayed dates could vary
when hovering them multiple times.

This is easily fixed by copying the affected variable by value.